python-app/template/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
prometheus_flask_exporter
gunicorn